Renamed communication  to message
