+++ src/config/RouterConfig.ts
@@ -5,8 +5,9 @@ import youTubeChannelActivityControllerCB from '@controller/YouTubeUploadsContro
 import youTubeVideoInfoControllerCB from '@controller/YouTubeVideoInfoController'
 import statusControllerCB from '@controller/StatusController'
 import { createJwtControllerCB } from '@controller/JWTController'
-import { addCommunicationController, getCommunicationController } from '@controller/CommunicationController'
+import { addCommunicationController, getCommunicationController } from '@controller/MessageController'
 import apiKeyAuthenticationMiddleware from '@middleware/APIKeyAuthentication'
+import { getEventsControllerCB } from '@controller/EventsController'
 
 export default class Routes {
 	static BASE_URI = '/api/v1'
@@ -24,7 +25,11 @@ export default class Routes {
 		app.get(`${Routes.YT_FUNCTIONALITY_BASE_URI}/video/giveaway`, validateJWTMiddleware, youTubeGiveAwayControllerCB)
 		app.get(`${Routes.BASE_URI}/auth/jwt`, apiKeyAuthenticationMiddleware,  createJwtControllerCB)
 
-		app.get(`${Routes.BASE_URI}/communication`, getCommunicationController)
-		app.put(`${Routes.BASE_URI}/communication`, validateJWTMiddleware, addCommunicationController)
+		// message endpoints
+		app.get(`${Routes.BASE_URI}/message`, getCommunicationController)
+		app.put(`${Routes.BASE_URI}/message`, validateJWTMiddleware, addCommunicationController)
+
+		// events endpoints
+		app.get(`${Routes.BASE_URI}/events`, getEventsControllerCB)
 	}
 }
